Fixed several compiler warnings for Visual Studio.
git-svn-id: svn+ssh://svn.code.sf.net/p/supertuxkart/code/trunk/supertuxkart@1340 178a84e3-b1eb-0310-8ba1-8eac791a3b58

+++ src/gui/base_gui.cpp
@@ -129,7 +129,7 @@ BaseGUI::update(float dt)
 
 //-----------------------------------------------------------------------------
 void
-BaseGUI::TimeToString(const float TIME, char *s)
+BaseGUI::TimeToString(const double TIME, char *s)
 {
     int min     = (int) floor ( TIME / 60.0 ) ;
     int sec     = (int) floor ( TIME - (double) ( 60 * min ) ) ;
